Added colors.
- Now every tile has a back and a fore color.
- Now every actor has its own color.
- Moved core game-loop functions from `main.c` into `game.c`.
