+++ include/game.h
@@ -1,0 +1,11 @@
+#ifndef GAME_H
+#define GAME_H
+
+#define SCREEN_WIDTH 80
+#define SCREEN_HEIGHT 60
+
+void init();
+void update();
+void end();
+
+#endif // GAME_H
