+++ include/symbols.h
@@ -1,9 +1,18 @@
 #ifndef SYMBOLS_H
 #define SYMBOLS_H
 
-#define PLAYER 'g'
+#include <libtcod.h>
+
+#define PLAYER '@'
 #define WALL '#'
 #define FLOOR ' '
 #define ENEMY '%'
+
+#define FLOOR_BACK_COLOR (TCOD_color_t){0, 0, 0}
+#define FLOOR_FORE_COLOR (TCOD_color_t){91, 71, 51}
+#define WALL_BACK_COLOR (TCOD_color_t){0, 0, 0}
+#define WALL_FORE_COLOR (TCOD_color_t){79, 73, 67}
+#define PLAYER_COLOR (TCOD_color_t){204, 194, 184}
+#define ENEMY_COLOR (TCOD_color_t){137, 29, 29}
 
 #endif // SYMBOLS_H
