+++ src/CMakeLists.txt
@@ -1,1 +1,1 @@
-target_sources(rlike PRIVATE main.c map.c actor.c vec2.c vec.c event.c gui.c)
+target_sources(rlike PRIVATE main.c map.c actor.c vec2.c vec.c event.c gui.c game.c)
